+++ src/muncher/feats/feats.js
@@ -24,7 +24,7 @@ const FEAT_TEMPLATE = {
 function buildBase(data) {
   let result = JSON.parse(JSON.stringify(FEAT_TEMPLATE));
 
-  result.name = (data.fullName) ? data.fullName : data.name;
+  result.name = data.name;
   result.data.description.value += `${data.description}\n\n`;
   result.data.description.chat += `${data.snippet}\n\n`;
 
@@ -36,7 +36,7 @@ function buildBase(data) {
   if (data.prerequisites.length > 0) {
     const requirements = data.prerequisites.map((requirement) => requirement.description);
     result.data.requirements = requirements.join(", ");
-    result.data.description.value += `h3>Requirements</h3>\n\n<${requirements.join("\n\n")}\n\n`;
+    result.data.description.value += `<h3>Requirements</h3>\n\n${requirements.join("\n\n")}\n\n`;
   }
 
   result.data.source = utils.parseSource(data);