+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Improved senses parsing. Blind Fighting now imports blindsight.
 * Enriched action parsing for things like Unarmed Fighting actions.
+* Fix feat import requirements section.
 
 # 0.2.23